Folders filewatcher events

+++ src/MapTileEditor3D/FileSystem.cpp
@@ -45,7 +45,7 @@ std::string FileSystem::OpenTextFile(const char* path)
         {
             std::string contents;
             in.seekg(0, std::ios::end);
-            contents.resize(in.tellg());
+            contents.resize((const unsigned int)in.tellg());
             in.seekg(0, std::ios::beg);
             in.read(&contents[0], contents.size());
             in.close();
